requirements.txt
pandas~=2.3.3
streamlit~=1.53.0
python-dotenv~=1.2.1
openai~=2.14.0